test(added tests of DialogComponent): dialog
