+++ package.json
@@ -12,7 +12,7 @@
   "config": {},
   "size-limit": [
     {
-      "limit": "15 KB",
+      "limit": "20 KB",
       "webpack": false,
       "path": "dist/**/!(*.test).js"
     }
