+++ src/components/Dialog/DialogTitle/_stories_/DialogTitle.stories.tsx
@@ -1,4 +1,4 @@
-import { DialogActionsProps } from '@bluebase/components';
+import { DialogTitleProps } from '@bluebase/components';
 import React from 'react';
 import { getComponent } from '@bluebase/core';
 import storiesOf from '@bluebase/storybook-addon';
@@ -6,7 +6,7 @@ import { withInfo } from '@storybook/addon-info';
 import { withKnobs } from '@storybook/addon-knobs';
 
 
-const DialogTitle = getComponent<DialogActionsProps>('DialogTitle');
+const DialogTitle = getComponent<DialogTitleProps>('DialogTitle');
 
 const stories = storiesOf('DialogTitle', module);
