+++ src/components/Dialog/DialogTitle/index.tsx
@@ -5,10 +5,10 @@ import React from 'react';
 const DialogTitle = (props: any) => {
 
 	return (
-    <MuiDialogTitle  {...props}>
-      {props.children}
-    </MuiDialogTitle>
-  );
+		<MuiDialogTitle  {...props}>
+			{props.children}
+		</MuiDialogTitle>
+	);
 
 };
 export { DialogTitle };
